mypage api 수정
